Menu improvements:
* Settings, help and about buttons are removed from the nav drawer when running in electron.
* Menu items for Settings, help and about now open the app dialogs properly.
* Windows and linux have 'Exit' and 'Settings' added to the 'File' menu.
* Windows and linux have 'About' added in the 'Help' menu

The menu declaration is getting... unruly. I think in the future we'd want to abstract it out and also add support for localization. But for now this will have to do.

Fixes #182

Moved isElectron method to be static.

Fixed analytics by including a local copy of the tracks code.

Wasn't working because of relative Urls that were included in the `w.js` file that was downloaded from wp.com.

Revert "Fixed analytics by including a local copy of the tracks code."

This reverts commit 3fa216330ae93cd3bb77350f1e7a79ae5bc71288.