Resolves #212
Previously the app would allow sending an authentication attempt with no
values written in the username or password fields. This returned a valid
response, but the app didn't interpret that response properly and thus
entered an authenticated state without having a valid login.

Technically, the response came back as valid JSON, but the
`access_token` property was undefined, which got saved in `localStorage`
as `'undefined'`, a value that failed to test as empty but which in
reality was.

If an app is currently in this undefined state, it might need to have
its localStorage reset. This not should occur moving forward since the
guards are in the authentication function.

 - Prevents sending the auth request with empty username or password
 - Placed multiple guards against using invalid login states