Removed custom react-hot-loading devServer for webpack-dev-server
The hot reloader was adding messy `ProxyComponent` wrappers around
everything and added unecessary complexity to the build process. This
commit removes that in favor of a config-less `webpack-dev-server` tool
instead.

This may lose some fluidity in propagating changes, but
webpack-dev-server now support auto-reloading anyway, so this new way is
probably good enough.