Strips `font-weight` property for note title in list
Resolves #107

Previously Chromium was glitching with `font-weight: bold` when trying
to show Emoji. Instead of the Emoji, it showed nothing.

An attempt was made to dynamically replace the Emoji with short spans
with priority on `font-weight: normal` but performance was an issue.

Since the trunk build of Chromium has resolved this and another
Emoji-related bug, this patch removes the bold property for now to get
around the problem.