Hide publish link if note not published
Resolves #130

Previously in the share dialog, the published link field would appear
even if the note wasn't being shared.

Now, the pushlished link field only appears if the published toggle is
on. Additionally, the copy link only enables if a shortlink is
available, but not during the waiting time to get that link.