Changed note-detail to handle the saving of the note itself.
The note will be saved if 3 seconds have passed since the user stopped typing, or if a different note is selected.

Utilized debouncer and its flush method to make the code much cleaner.

Hattip @dmsnell

Remove debounce module variable in favor or local state

Previously the debounced note save function was stored in the module
state, which opened up the opportunity for bugs occurring when doing
things such as creating multiple copies of the note detail.

This patch moves it inside the React component, with the small caveat
that it has to be created in `componentWillMount` due to the dependence
on `this.saveNote`. Right now there is no default, so maybe it would be
safer to provide a `noop` for before `componentWillMount()`, but I'm not
sure that would matter because of that lifecycle method being called
before the component instantiates.

Updates to lodash bump

Renaming queueNoteSync to queueNoteSave