WIP: Replace confusing app state with simpler Redux store
The existing app store and reducers have been pretty complicated and
confusing for me to figure out. They are definitely inspired by a Redux
model, but have taken the things that are great about Redux and spoiled
them: action creators that are just simple objects, stateless reducers,
composability.

This bigger PR is attempting to replace all of the custom code with
simpler reducers and a more proper use of the `connect()` function to
prevent having to pass so many props down the component chain.