Hook two-way communication into Electron main app to rebuild menu
The menu appears to have previously only had one-way communication,
which felt a bit embarassing since we didn't even display a checkmark
for which theme was selected.

This lack of communication also made fixing things like #236 difficult
because the menu needs to respond to events that happen within the app.

This patch introduces a new signal that comes from the rendered window
into the main Electron process which passes the settings whenever they
have been updated. Note that since the settings are now updated with a
simple Redux flow, it's easy to reason about them and the fact that they
are immutable, giving us confidence that we won't rerender the menus
willy-nilly.

As part of this change, I started using `electron.ipcRenderer` and
`electron.ipcMain` instead of `ipc` by itself. Although I read some
documentation suggesting that we should be doing this anyway, I was
having quite a bit of trouble getting it to work without the change.