Cleanup docs/packaging.md
Removes "brew install" for Linux, we don't use Brew in any distro, this is for
OS X.  This also removes references to tar which is generally included in most
base Linux because of other dependencies.  It also adds more ways to install
Ruby because not everybody wants RVM in their home directory and also provides a
way for users to install fpm with bundler.