Pressing enter logs in user now.
This fixed issue #248.

Enter key bindings were missing on auth.jsx. I added onKeyDown props to
both of the inputs to handle the case of if user enters password, or case of
typing invalid credentials.

In onLogin method added event type check to determine if the event is
keypress or mouseclick. If it is a keypress and it is not Enter key,
just keeps continue regular event handling.
I placed check before preventDefault to make sure the event handling works as
expected on other keypresses than Enter key.