Inserts a tab character in the editor on 'tab' press
Previously, hitting the tab key while in the note editor would advance
the keyboard focus to tag list input but would not insert a tab
character in the note content.

Now, the tab character is inserted and the keyboard focus is not
advanced.

This inserts the tab character just fine, however it does not allow for
an "undo" of the insert operation. This should be fixed prior to
merging.