Handle tab and shift-tab
In earlier commits on this PR we were simply inserting a tab when the
tab key was pressed, but this left an awkward experience when using
shift+tab.

Now, when we have multiple lines selected, tab and shift-tab increase
and decrease the indentation level respectively according to the
expected behavior.