Automatically add list bullets on new list items
Previously, there was a missing behavior in this app that was present in
the other Simplenote apps: when creating a new line after a line
starting with a list bullet, the new line should start with the same
bullet at the same level of indentation.

This PR adds that behavior into the app.