Autocomplete tags as you type
Change representation of selected tag

In the note editor we can select a tag before deleting it. This is a
confirmation mechanism to prevent accidental tag deletion.

Previously this "selected" tag was being stored as a reference to the
position in the tag list holding the specified tag. This system was
complicated and prone to error because when the tag list changed
externally, the index was no longer valid.

However, the tags can be entirely specified by the same string that
displays when showing the tag and is a sufficient reference in the code.

This patch stores the selected tag by its own name. By doing this we
obviate the need to rearrange references and shuffle around and it also
prevents storing the "selected" note index as `-1`, which is an
impossible index into the list.

Single-click on tag to delete it