Finds note from `state.notes` instead of using `state.note`
When `onNoteUpdate` was being called it was looking for `state.note` but
that didn't exist in at least some calls. I have not fully understood
the `state.note` variable or when it should exist.

In this patch, I have manually scanned the stores list of notes to try
and find it, working around the need for `state.note` to be set.

This is fixing some exceptions that were raised. I am not sure how many
other issues it will resolve but it was preventing the call to
`updateNoteContent`