Fix revision slider
Previously there was an issue with the revision slider in that the items
were out of order chronilogically and henceforth the 'Latest' revision
was being placed in position 1.

This update introduces a sort on the revisions by `modificationDate` to
force order and updates the state of the slider as revisions loads in
order to make sure that the latest revision is on the far right and
automatically selected if no other revision is selected.