Fixes search bug and refactors `filterNotes()`
Previously in #382 we replaced the `RegExp`-based note search with a
simple string-based search.

In that PR we apparently missed the fact that we were rejecting notes
whose content matched the search query instead of rejecting the ones
whose content _didn't_ match the search query.

This has been fixed.

Alongside the fix I have refactored the function to be clearer in hopes
of making future bugs like this less likely.