Clear search field on ESC keypress (#416)
This patch adds the behaviors where pressing the _escape_ key while in
the note search box will clear out any text already filled in.

This is a precursor to adding the clear button as requested in #411.

In order to make this work properly I had to make the search field
itself a controlled component whereas before the `value` was free as any
given HTML input would be. Now it's tied to the `filter` app state
property.

I have also taken the opportunity in this patch to refactor the
`<SearchField />` component, converting it to a `class`-based React
component and wrapping it in `connect()`, providing whichever props I
could without major app refactoring.