Set initial filter (search query) state to '' (#417)
Previously the initial filter on the notes was undefined until someone
actually typed something into the search field.

Now, the filter is set as an empty string on boot. This not only
provides us with the inital filter required by the search field but it
also keeps the filter state in line with the expected type of data it
needs to be: a string.