Redux: Move authentication state into simple Redux state (#409)
* Move Redux code into new `state` directory

This didn't belong in the `flux` directory where it was previously
located. Instead, it has been moved into `state` to better reflect that
this will be the place where the app data lives.

The settings files have been renamed to reflect standardizing state
sub-trees.

* Refactor auth state out of appState and into Redux proper

* Review feedback: bugs

* Provide support for login errors

Previously there was no mention of having a login error. Because of
this, the "Credentials don't match" message was appearing whenever
someone was logged out.

Now there are explicit cases to handle logged out, invalid login
attempts, and unknown failures to login.

* Update some code for terser style