Use cross-platform shortcuts for menu accelerators
Resolves #378

Previously we were using `Command+` shortcuts in some menu accelerators.
This works on Mac computers but fails on Windows and Linux computers
(whre the app actually runs).

Now these have been replaced with `CommandOrControl+` shortcuts which
still mean that on Mac computers it will catch on the Command key but on
the other platforms it will catch on the Control key. See [the Electron
docs](http://electron.atom.io/docs/api/accelerator/#platform-notice) for
clarification.