Add hidden export function (part 1)
Addresses #332
Addresses #323

The Simplenote app receives updates from the Simperium library
(`node-simperium`) and stores those updates inside of `IndexedDB`. This
provides us with a fairly reliable local copy of the note data, but that
data is not very accessible from outside of the app.

The web app on simplenote.com offers a `.zip` archives of an account's
notes with one file per note but the client applications do not offer
any kind of an export.

This patch introduces an export function into the Electron app which can
be called to export a single _JSON_ file with all the relevant note
information. This export contains metadata unavailable from the existing
export on the web.

There is not piece of this built into the user interface yet and so this
PR introduces a feature for testing and development only. The export
function is made available in the global scope and can be called from
the running console as `exportNotes()`. It should trigger a download of
the `notes.json` file.

All notes are being exported but trashed notes and normal "active" notes
are separated as two different lists.