Fix: Lists should be prefixable by `+` bullets
In #398 @nfcampos added the ability to autocomplete lists. However, only
`-` and `*` were chosen as valid list bullets. This doesn't match the
behavior in the other Simplenote apps and it's not our desired list
behavior.

This PR adds `+` as an option and in the process extracts some of that
validation into a helper function.

Additionally it resolves a small bug where space-prefixed bullets
weren't registering as lists.