Create new `revision` Redux state subtree and remove `NoteEditor` internal state
The new `revision` subtree is used to hold a note’s revision and revision selector display states.
It replaces the `<NoteEditor>` internal state, hence fixing the weirdness of it being controlled by children components (`<NoteToolbar>` and `<RevisionSelector>`).

The `revision` state is intentionally simple and in the future it will probably be replaced by a more complex `revisions` subtree, which is definitely out of this PR scope.