Change method for showing autocompleted text
Previously we were hiding an `<input>` underneath the actual tag
`<input>` and displaying the autocomple suggestions in that underlying
field. This caused styling complexity and errors when the suggestion
text didn't match the actual text in case: for example, when the user
types `re` and the autocomplete suggests `Recipe`. The differently-cased
letters would show a visual glitch.

Now we're using two `<div>`s side-by-side with the first being a
`contentEditable`. The `contentEditable` is now the tag entry and the
second shows the suggestion minus the actually entered prefix. This has
worked significantly better and is simpler than the stacked `<inputs>`.

Finally this updates the styling to look right.