Prevent creating duplicate tags having different cases
Previously it was possible to create a new tag of the same name as an
existing tag if the casing were different. For example, if the `Recipe`
tag already existed it was possible to create the `recipe` tag because
the first letter was different.

Now this has been prevented by performing a case-insensitive list union
on the old and new tags before updating the note objects.

Additionally this patch suppresses the React warning about using a
`contentEditable` as a controlled components.