Autocomplete tag names when adding new tags (#389)
**Change representation of selected tag**

Previously this "selected" tag was being stored as a reference to the
position in the tag list holding the specified tag. This system was
complicated and prone to error because when the tag list changed
externally, the index was no longer valid.

However, the tags can be entirely specified by the same string that
displays when showing the tag and is a sufficient reference in the code.

This patch stores the selected tag by its own name. By doing this we
obviate the need to rearrange references and shuffle around and it also
prevents storing the "selected" note index as `-1`, which is an
impossible index into the list.

**Auto-delete tags on click**

In the note editor we been able to select a tag before deleting it. This is a
confirmation mechanism to prevent accidental tag deletion. However, that safety isn't providing too much value compared to it pushing in one more step before deleting tags. Tag changes should be saved in the note history as well so if we accidentally delete a tag we can revert the change.

**Prevent creating duplicate tags having different cases**

Previously it was possible to create a new tag of the same name as an
existing tag if the casing were different. For example, if the `Recipe`
tag already existed it was possible to create the `recipe` tag because
the first letter was different.

Now this has been prevented by performing a case-insensitive list union
on the old and new tags before updating the note objects.