Add Export (to zip) to app when running in Electron (#448)
Addresses #332
Addresses #323

The Simplenote app receives updates from the Simperium library
(`node-simperium`) and stores those updates inside of `IndexedDB`. This
provides us with a fairly reliable local copy of the note data, but that
data is not very accessible from outside of the app.

The web app on simplenote.com offers a `.zip` archives of an account's
notes with one file per note but the client applications do not offer
any kind of an export.

This patch introduces an export function into the Electron app which can
be called to export a file with all the relevant note information.
This export contains metadata unavailable from the existing
export on the web.

This patch adds the ability to export the notes as a .zip in the same
way that this behavior works on app.simplenote.com

First we generate the JSON export and then process that data to build
the zip archive. The **JSZip** library does the hard work.