Fixes a bug where the note list was shrinking for notes with many tags
Previously when selecting a note with lots of tags the note list would
shrink to a narrow width.

This has now been prevented by disabling the auto-width setting of the
`<AutoSizer>` wrapper in the note list.

We may be able to remove the `<AutoSizer />` if we can set a fixed
height but this is sufficient for now and it doesn't appear to be
costing much in terms of performance if at all.