Fixes a bug in the logout flow where the app becomes unresponsive (#493)
* Fixes a bug in the logout flow where the app becomes unresponsive

Previously but after some auth state structure, when logging out of an
account, the app would become unresponsive. The cause of this was an
error raised in the tag input component which was attempting to add an
event listener to a component which was no longer mounted.

This fixes the symptom and may fix the root cause but at this time I am
not certain what that root cause it. At this time I believe that the
`ref` function is being called on component unmount which I did not
anticipate, passing in `null` and thus failing to add the event
listener.