Add command-line option to open devtools
Previously one needed to change the source code of the app in order to
open the developer tools. This also meant that one needed to undo those
changes before checking in and committing code.

Now, we can pass in a command-line option after the app has been built
in order to determine if the tools open or not. This means no risk of
accidentally merging in changes which open the tools by default and it
means we don't have to recompile in order to inspect the running app
inside of Electron.