Fix: Can't navigate to note list on small widths without resize
When starting the app in a narrow-width state it can be impossible to
navigate back to the list of notes unles a manual resize of the window
is performed.

The reason for this behavior appears to be improperly initialized state
from the browser window in the `browserShell` wrapper. Instead of
grabbing the width attributes from the window when initializing we were
setting default values of 1024 px and a wide screen. In turn, these
values remained until a resize corrected them, therefore prohibiting the
navigation.

The values are now initialized directly from `window` and the
`browserShell` higher-order component's style has been updated from
`React.createClass()` to return a `React.Component`.