Add/copy rendered markdown (#579)
* Markdown: Replace `marked` with `showdown` parser/printer

`marked` doesn't appear to have been maintained for a while while
`showdown` has more recent contributions. Further, `marked` was missing
some functionality that we wanted. Namely, we wanted an equivalence with
Github-flavored Markdown (gfm) and `showdown` does it better.

In this patch we've changed the way we are inserting the content.
Instead of using `dangerouslySetInnerHTML` we're statefully rendering
the content into the DOM and then manipulating it. This change provides
an easier way to interact with the rendered content because we can
operate at the DOM level instead of at the string level. For example,
this is how highlighting now gets inserted.

**Notable changes**
 - Todo items render (though they can only be changed in the plaintext
   view)
 - Tables render
 - Code blocks highlight based on their language
 - Certain dirty/malicious content gets sanitized before display

**Questions**
 - Since the todo items render but aren't editable it could be confusing
   for someone to use, misleading them into thinking it's supposed to
   work

* Markdown: Add ability to copy rendered note HTML

When viewing a note's Markdown preview and with no selection, if you
press the copy shortcut then the rendered HTML for the note will fill
the clipboard buffer. If there is some selected text then the copy will
behave as normal.

* update dll config