Update build environment versions - webpack1 > webpack3
This change is mainly about updating `webpack` from the 1.x versions to
3.x versions. Several secondary changes are the result of that.

The app is now being built with es2015 modules instead of CommonJS
modules. Theoretically this should shrink the build size and maybe speed
things up on account of webpack's tree-shaking algorithm and having
fewer closures in the final code. I haven't tested yet and I don't
anticipate this being significant as an Electron release.

Several small updates are going on too:
  - The node version has been bumped to 7.9.0 to match that running
    inside of Electron.
  - `package.json` doesn't need `npm bin` to prefix commands
  - I'm trying to let Babel focus on the version of Chrome that Electron
    is running and ignore older browsers.
  - We weren't using `AppCache` and didn't need it. It's a deprecated
    API anyway.