Release: Require `appdmg` to dependencies (#636)
In order to build the macOS bundle we need to be able to create the
`dmg` file. Unfortunately this package won't install properly on
non-macOS machines. This patch adds the dependency to
`optionalDependencies` which should not prevent the `npm install`
process from continuing if it fails. For macOS builds it should install
and for others it should skip without causing trouble.