Markdown: Interactive preview mode switching implicity
In all of the simplenote apps there has been an inherent distinction
between edit mode and preview mode when interacting with
Markdown-enabled notes.

In this patch we're exploring the implicit transitions, such that the
default view for a note is the rendered markdown "preview" but as soon
as you click to edit the note if flips back into the edit view.

This was inspired by prior work and for now is an experimental change.