Editor: Replace draft-js with Monaco
We've had a few issues with draft-js including performance conderns.
Recently we've also been wanting to explore Markdown syntax highlighting
and while draft-js is ammenable to that it leaves considerable
implementation details up to us.

In #568 I proposed a change to replace draft-js with a plain old
textarea which greatly improved the performance for large notes but also
had the major drawback that we lost our ability to easily decorate text.

Microsoft's Monaco editor powers Visual Studio code and is incredibly
performant thanks to the fact that they have rebuild all the editor
behaviors in custom ways to handle big files and syntax highlighting.

In this patch we're taking advantage of the work that Microsoft has done
and I'm proposing a switch to Monaco. It includes Markdown syntax
highlighting for free. This comes through a React wrapper around the
editor.

The only real drawback I see here is that Monaco is much larger than the
textarea in terms of raw KB though I'm not sure how it compares to
draft-js. In Electron this isn't too much of a concern but if/when we
push this version of the app out to the web it could come with other
problems. Monaco also doesn't officially support mobile devices either I
think.

**Remaining work**

Monaco already has great auto-indent support and provides many of the
behaviors for us which we were implementing by hand with draft-js or in
the textarea. However, it doesn't provide all such as auto-lists.

There remains some dataflow issues that exist already in the app too but
which aren't solved here. When making changes it's possible to queue up
changes but get updates from the server in the meantime and mess up the
editor.