Update to React 16
This major upgrade involves only a few basic changes:

 - No more `React.createClass()`
 - `PropTypes` from separate npm package
 - `onClickOutside` is now a higher-order component instead of a mixin

In this patch I've updated the necessary dependencies and made the
necessary code changes in order to bump our React version.

Performance _should_ be better but probably not noticeably so.
This update is more to keep in line with React updates.