Detach editor from state; add autolist
In the process of adding the auto-list functionality I had to disconnct
the editor from state and take it from controlled mode to self-contained
mode. This means that external updates won't come in.

I had to do this because I realized that whenever changes came in from
the parent component it would wipe out the undo stack.

The solution which should come is to intercept the updates coming in
from the outside and manually mirror those in the editor. This has been
what we've needed for a while actually and it could solve other issues,
but it's going to take a while to get right probably.