Update editor on external changes
In this patch I was able to update the contents of the editor when
external changes arrive and I was even able to update the cursor
position. However there is still one major issue here: a race and
some unexpected data flows when conflicts exist.

In every test I tried out I was unable to find a transformed patch where
one should have rebased. In fact, I wasn't able to find any situation in
which the working patch wasn't undefined.

This is unexpected because I made changes in the working editor after
changes in another window were queued and that meant that the working
copy _had_ to be different.

Alas, during conflicting edits the content is mangled.

This may exist on **master** (and I will confirm).

It has been my running suspicion that our `queueNoteSave()` is to blame.
It currently waits two seconds before submitting changes.  If new
updates arive during that two second window then the local props will
have changed underneath it. Because it hasn't submitted the local copy
to the internal Simperium client by the time the updates arrive then
there's nothing to transform. Because it submits the local copy
(captured in the debounced-function closure) _after_ the external
updates have already been applied then Simperium must think that we not
only wanted to mak our changes but also undo the external updates.

I will try and reproduce and explore in **master**.