Note Updates: Resequence async calls to match execution flow
The `noteUpdateRemotely` action will attempt to update a note's contents
in the `noteBucket` if there are changes in the local copy that need to
rebase against updates from the remote server.

The code as written has made it appear as though this update occurs
before calling `loadAndSelectNote` but because of the asynchronous
behavior of the `noteBucket` the update occurs after that action is
dispatched. The `loadAndSelectNote` action further complicates things by
calling `noteBucket.get()` which is itself asynchronous and so ordering
of events is even more dubious.

In this change I'm rearranging some of the control flow and performing
an early-abort in the case where we have no rebasing to apply. This
should hopefully make it slightly clearer in what sequence the events
are occurring.

Likely this will have no impact on the application: the change is mostly
duplicating some code in a way which shouldn't actually change the
runtime operation.