Deps: Remove redundant `md5` library
See #1308

When we added a new direct dependency on `md5` we inadvertendly created
three new dependencies to the application. We intended to _remove_ a
depdendency on `create-hash` but that library was already transitively
pulled in via several other packages. Those packages and `create-hash`
depend on `md5.js` and when we added `md5` as a requirement it added
something we didn't already have.

In this patch we're replacing our use of `md5` with `md5.js` and thus
removing the added dependencies, since `md5.js` is already in the
project. An alternative approach to remove the excess dependencies would
be to revert #1308 and rely directly on `create-hash` again.

Before applying this patch there are 1160 directories in `node_modules`.
After applying this patch there are 1157 directories in `node_modules`.
This count may or may not be relevant.