Build: stop failing the build on eslint errors
`eslint` provides helpful warnings that prevent bugs in our software.
However by running it as a precondition for the build we introduce
artificial obstacles to quick testing and iteration. For instance, we
will fail to build the app if we have an unused variable even though
there's no reason it would break the build.

In this change we're removing the precondition for passing `eslint`
before building the app. This should allow us to more rapidly made small
debugging/testing adjustments while working on fixing or altering the
app.

`eslint` still belongs in the process and we might follow-up on this PR
by introducing it as a `git` hook on `precommit`, a more appropriate
place for the safety check. For now though I am foregoing that work in
order to remove what has been a persistent frustration while working on
the app.