This PR changes the way that the editor's contents and Simperium interact; it's dependent on the changes made in simperium@1.0.0.
Previously we held content in the editor component and would delay
sending those changes to the underlying Simperium client library
(node-simperium) until the timeout flushed. This made for awkward
interactions when we would received remote updates from the server that
happened before that timeout flushed; this led to note corruption.

After this PR we are announcing immediately to the underlying client
library when we make changes and telling it (via { sync: false } in
noteBucket.update()) that it's okay to wait to synchronize those changes
with the server. We still maintain our debounce, but instead of sending
the editor component's text on the timeout we more simply tell the
underlying library to flush its changes (via noteBucket.touch()). The
effect is that we've removed a source of concurrency bugs with those
remote updates.

After this PR we should be removing a number of related data loss and
note corruption issues all relating to the interplay of queued and
received changes.