Fix: Stop rebasing updates after they come in
See #1579, #1598
See Simperium/node-simperium#78
See Simperium/node-simperium#61

When we receive an update from a remote client we have been listening
for it and adjusting our local app state to account for that change.
Unfortunately in cases where we also have local or pending changes we
have been erroneously transforming or _rebasing_ the change before
applying it. The underlying `node-simperium` library has already
performd that transform, however, and when the client application
does that too it leads to double-writes and misapplied patches,
producing note corruption.

In this change we're stopping the rebase operation we have been
performing and that will remove this particular bug from the
application. This change does not solve all of the problem however
because we also have to make sure that the `node-simperium` library is
aware of our local state when it receives those updates.

Further work is taking place in Simperium/node-simperium#61 and #1598 to
close the loop on that but these changes are important enough on their
own to warrant a change.

This will close one bug while opening another but this is a dependent
part of the process in closing the broader issue.