Fix: Sort revisions by note version
See #394 where the sorting was introduced

When we make changes to a note we append a `modificationDate` value
which serves as a timestamp of the change _on the client device making
the change_. That is, the timestamp is meant to provide an indicator of
when in real time the change happened, but since it's based on a
device's clock we cannot rely on it or use it for proper sequencing.

It happens from time to time (especially when one device is offline)
that a revision which occurs _after_ another remote change gets a
timestamp showing that it happened _before_ because the local device
recorded the timestamp before the change made it to the Simperium
server.

When this happens the application will show those revisions out of the
actual order they took place and this can lead to confusing note
history. In this patch we're sorting the revisions based on the version
of the note and not the timestamp. This produces the proper sequencing
of changes in the note's history.

This change will present certain scenarios where it appears as though
the revisions are out-of-order because they are showing their
modification times even when those conflict with the version-ordering.
We should consider a way to communicate this, possibly with a tooltip
or a contextual notice when we detect these conflicts.