Refactor `pinNote` to stop directly mutating note object
Previously we have been directly mutating the `note` object in the
action creator when toggling a note's pinned status.

This mutation can lead to concurrency defects which expose themselves as
inconsistent UI state. This breaks our Redux model which assumes that
all UI updates happen atomically.

In this patch we're building a new `note` object when toggling the
pinned status in order to maintain our consistency.

At the same time I have removed the thunk from the action creator since
it wasn't doing anything and didn't need to exist. This shouldn't have
a noticable impact on the application.